se cambio el metodo eliminar de la tabla plato

+++ Personal/Admin/Platos-Eliminar-neg.php
@@ -1,19 +1,24 @@
 <?php
-
 	session_start();
 
 	include ('conexion.php');
 
-	$idPlato = $_REQUEST['idPlato'];
+	if(isset($_POST["idPlato"])){
 
-	$sql = "DELETE FROM Plato where idPlato = '$idPlato'";
-	if (!$result = $db->query($sql))
-	{
-	    die('No hace consulta a Usuarios ['.$db->error.']');
+		$idPlato = $_POST['idPlato'];
+
+		$sql = "DELETE FROM Plato where idPlato = '$idPlato'";
+		if (!$result = $db->query($sql))
+		{
+		    die('No hace consulta a Usuarios ['.$db->error.']');
+		    $response['status'] = '2';
+            echo json_encode($response);
+		}else{
+			$response['status'] = '1';
+            echo json_encode($response);
+		}
+
+	}else{
+		echo "SE ENTRO DIRECTO A ESTA PAGINA";
 	}
-
-	$_SESSION["Plato_eliminar_ok"] = "";
-
-    header ("location: Platos-Crear.php");
-
 ?>
